Rename uuid module to bleuuid to avoid clash with uuid crate.
